Provide ALIASES and NAMES table.
The ALIASES mapping can be used to specify aliases for names
including state point name and values. That makes it easier to
use fully qualified names for state point definition, but still
have a compact overview.

The NAMES mapping can be used to control certain names in the status
output, e.g., for translation.