The scheduler jobs map handles multiple scheduler job per job id and op.
The scheduler jobs map now contains a list of scheduler jobs associated with
one job id and operation name. Previously the map was constructed with the
assumption that no more than one scheduler job may be associated with one
pair of job id and operation name. However, in reality this may be possible,
for example when one scheduler job just stopped, and the next one is already
queued up.