Make update_status run faster
A __setitem__ on a job document triggers a re-read and then a write of
the json file. These I/O operations are very slow on scratch filesystems
with slow metadata servers.

Avoiding triggering a __setitem__ call when it is not necessary.
This change significantly increases the performance of
update_stati().