Implement the directives operator function decorator.
The directives decorator is intended to be used to provide additional
execution directives for a given operation. A prominent example is to
specify the number of required processors.

The directies dictionary is converted into a namespace and is made
accessible to the templating context. This can then be used by the
template implementations to calculate required resources and respond to
other directives.