If necessary, initialize a signac project when calling 'flow init'.
The signac project will only be initialized if a call to
signac.get_project() would fail.

Note: The alias provided to flow init must be a valid Python identifier,
so the project name is also restricted in that case. In general, signac
does not limit project names to valid Python identifiers.