Merged in revise-verbosity-options (pull request #27)
Revise verbosity options

Revise the command line options for logging and traceback verbosity.

This patch adds the -v/--verbose and --show-traceback arguments to
the command line interface. It enables users to control the output logging
verbosity and whether the traceback is shown with two separate options.

The --debug option now implies '-vvv --show-traceback', which
replicates previous behavior.

In addition, this patch fixes an issue where these global options would
need to be provided in front of the sub command. The global options
should now be recognized anywhere within the command line string.

Approved-by: Vyas Ramasubramani <vramasub@umich.edu>
Approved-by: Carl Simon Adorf <csadorf@umich.edu>