Enable the definition of "aggregate" operations.
The main change introduced in this patch is the refactoring of the
JobOperation class into JobsOperation. That means that an operation
operating on only one job is a special case of a general operation, that
operates on multiple operations. In that sense *every* operation is an
aggregate operation, but the standard case is that it operates on only
one job.

This is achieved in detail by:

 * adding an "aggregate" parameter to the FlowOperation class which
   classifies a specific FlowOperation as an aggregate operation.
 * adding the `@flow.aggregate` decorator which labels an operation
   function to define an aggregate flow operation,
 * adding the `aggregate` parameter to the
   `FlowProject.add_operations()` method,
 * refactoring the `JobOperation` class into `JobsOperations`, which
   operates on one or more jobs by default. A job-operation with only
   one job is therefore a special case of an aggregate operation.

It might be possible to remove the need for an aggregate decorator, and
instead just inspect the function signature, but we can reevaluate that
at a later point.