Correctly determine the source file for wrapped functions.
This patch fixes an issue where the source file, e.g., the project.py
file in which an operation function is define in, will not be correctly
detected in case that the function was wrapped with some kind of
decorator.

For example:

    # project.py
    from util import my_wrapper

    @Project.operation
    @my_wrapper
    def hello(job):
        pass

In case that my_wrapper redefines the function in util, the source
code file was previously detected to be util.py instead of project.py.