Add dummy job-operation to fix issue #127
Per offline discussion with @csadorf, this commit fixes creates a dummy
job operation if a job has exactly 0 eligible operations to fix a bug
that was introduced in PR #111; the bug causes the default detailed
status view to not show jobs with no eligible operations, which changes
the previous behavior.

I've only tested this on a real-life project that only has 2 different
operations, but it seems to fix the bug.