Revert changes to status view templates
Per discussion on github and offline, the best way to proceed with this
(somewhat temporary) fix is to not touch the templates as there are
multiple paths to manage if you do that (e.g., different templates for
stacked and detailed view must be maintained separately but in
parallel). This commit removes the changes to the base_status template,
so that the old behavior is achieved but with the new templating system.
The only drawback is that when you print the status in with the detailed
argument, you will now get empty square brackets for jobs that don't
have eligible operations.