Show number of eligible jobs in status view (#135)
* Show number of eligible jobs in status view

This commit adds functionality to print the number of jobs that are
eligible for each operation during a status view. I added this to the
"overview" section of the view, since it is an overview-like thing to
print. This commit addresses issue #134.

* Update changelog

* Sort num of eligible jobs in status, add max arg

For printing the number of eligible jobs for each operation, they are now
sorted by the number of eligible jobs. This commit also adds a command
line option to status to control the maximum number of operations to
print.

* Add default number of operations to print in status

Use config value to set default