Add Directives.__repr__ and fix bug in FlowProject._register_groups
Directives.__repr__ is not a valid way in Python to recreate the object.
To be clear there is no single Python expression to do this necessarily.
If any values are not the defaults then we must manual set each key.

The bug invovled always setting the operation_directives for a group
which cause those default directives to be used instead of the ones
found in the singleton group