Prevent bug in configuration
If used the deprecated argument `no_parallelization` in `print_status`
before this commit, the configuration value in
`project.config['flow']['status_parallelization'] would be changed for
the rest of the session. Now this argument only effects the current
function call.

This can be removed later once `no_parallelization` is removed.