Fix finalize_np
Had to change the logic to only use nranks * omp_num_threads if np == 1
(the default value). Otherwise, we may be overriding an aggregated or
user set value. We cannot use the nranks * omp_num_threads logic in non
default cases because we have no way of knowing how the current value
was acheived.