Entrypoint and command prefix improvements. (#302)
Entrypoint paths are now autogenerated in `FlowGroup._determine_entrypoint`. We assume that the correct file is the one that the project is defined in.

signac run prefixes  are now generated in the `FlowGroup._create_run_operations` function. We use the `fork` directive to specify that an operation has a nontrivial prefix. `JobOperation` commands are no longer lazily evaluated as well, since an entrypoint always has a default. This means that the submission output that shows the expected operations to run in a given group shows the correct prefixes.

Co-authored-by: Brandon Butler <butlerbr@umich.edu>