Fix Stampede Submission with FlowGroups (#298)
Updates the `Stampede2` environment class to work with intra-bundle parallelization which was the only parallelization before groups (inter-group parallelization was removed before this commit). Uses an environmental variable to allow for communication of the offset for each signac process that get started in parallel within the script.

Also, fixes bug where the expected operations print out could be in any order in Python 3.5 since we used a `dict` rather than a `OrderedDict`. This comes from a previous commit.

Co-authored-by: Vyas <vramasub@umich.edu>