Change the submission ID for aggregate logic (#334)
Changes the submission id generation to prepare for aggregation. The `__str__` method of `_JobOperation` is also changed to better handle job aggregation. The unique id generated using a hash of all the `_JobOperation` object's data is still the same, so the uniqueness should not be in question; the only change is to the readable part of the id.

The change to the `str` implementation mimics NumPy, and the submission id favors brevity.