Evaluate directives in _JobOperation Constructor (#409)
* Revert "Revert "Evaluate directives in _JobOperation construction.""

This reverts commit 6d0420788454368458e07da63a50bb58b88c70e5.

Revert the revert. The original reason to revert the first commit is it
was accidentally pushed to master. Now I want the commit on a branch.

* Make JobOperation use _JobOperation.__init__.

* Evaluate directives in FlowGroup._get_submission_directives

I also removed the evaluation in _JobOperation since this should be
redundent now.

* Add tests for proper evaluation of directives

The new tests ensure that directives for both singleton and
multi-operation groups are fully evaluated when creating _JobOperations
or one of its subclasses.

* Add changelog entry for fixing directives evaluation

* Apply suggestions from code review

Co-authored-by: Bradley Dice <bdice@bradleydice.com>