Modify _JobOperation construction to not deepcopy directives
Deepcopying directives is expensive since it is not a pure data
structure. This commit just converts the directives to a pure data
structure (i.e. a ``dict``). In addition, we pass the user specified
directives (as opposed to the environment specified directives) to
``_JobOperation``. This should improve the performance of run and
submit.