Optimization: don't deepcopy directives (#421)
* Modify _JobOperation construction to not deepcopy directives

Deepcopying directives is expensive since it is not a pure data
structure. This commit just converts the directives to a pure data
structure (i.e. a ``dict``). In addition, we pass the user specified
directives (as opposed to the environment specified directives) to
``_JobOperation``. This should improve the performance of run and
submit.

* Only evaluate directives when unevaluated.

Adds a check to prevent the same directive from being evaluated twice.
This should speed up the evaluation of multi-operation groups.

* Add back (fixed) comment about user directives.

* Update docstring.

* Use dict constructor instead of dict comprehension.

Co-authored-by: Bradley Dice <bdice@bradleydice.com>