Add FlowProject.operation.with_directives decorator
The new decorator adds the appropriate directives along with registering
the function as an operation.

To accomplish this, we create a new class operation, that is specific to
a ``FlowProject`` class like the preconditions and postconditions. We
also use the ``__init__`` method as a decorator function, and only
interact with the class and not any instances. In fact, it is impossible
with the standard instantiation model to get an ``operation`` object.