Add operation.with_directives decorator (#502)
* Add FlowProject.operation.with_directives decorator

The new decorator adds the appropriate directives along with registering
the function as an operation.

To accomplish this, we create a new class operation, that is specific to
a ``FlowProject`` class like the preconditions and postconditions. We
also use the ``__init__`` method as a decorator function, and only
interact with the class and not any instances. In fact, it is impossible
with the standard instantiation model to get an ``operation`` object.

* Move documentation examples to use ``FlowProject.operation.with_directives""

* Return to using an instance for FlowProject.operation

The overwriting of __init__ does not work

* Rename FlowProject specific operation to OperationRegister

* Revert OperationRegister.with_directives to method

This is necessary since reverting to using an OperationRegister object
not class

* Fix test name typo directies -> directives

* Add tests to improve codecoverage

This is to make the code coverage tests pass

* Fix bug where user groups could overwrite operation names

* Deprecate @directives and remove from all but one test

Also removes @directives from defined FlowProjects for tests

* Apply suggestions from code review

Co-authored-by: Bradley Dice <bdice@bradleydice.com>

* Remove old tests on invalid operation func signatures

Co-authored-by: Bradley Dice <bdice@bradleydice.com>